lap_history: widen counter so the ring keeps cycling past lap 256
lap_history_count was a uint8_t with an explicit cap at 0xFF.  Once the
counter hit 255 it stuck there, leaving slot = 255 % 16 = 15 pinned for
every subsequent lap.  All laps from 256 onward overwrote slot 15;
slots 0..14 froze on whatever was there at saturation.  Endurance races
that cross 255 laps silently lost lap history and emitted stale rows in
the 0x56 garage reply.

Widen lap_history_count to uint32_t and drop the saturation guard so
the ring keeps cycling.  Also fix results.c to walk the ring with the
same wrap-aware start index as handlers.c h_load_setup; without it the
JSON laps[] reordered later laps in front of earlier ones for any car
with more than 16 completed laps.

+++ accd/state.h
@@ -218,7 +218,14 @@ struct CarRaceState {
 	 * always emitting split_count = 0.
 	 */
 	int32_t		lap_splits_ms[ACC_LAP_HISTORY][3];
-	uint8_t		lap_history_count;
+	/*
+	 * Monotonically increasing lap counter — used by handlers.c to
+	 * pick the ring slot, and by handshake.c / handlers.c / results.c
+	 * to compute the wrap-aware first-lap index.  uint32_t (not
+	 * uint8_t) so an endurance race with > 255 laps doesn't freeze
+	 * the ring with every later lap pinned to slot 255 % 16 = 15.
+	 */
+	uint32_t	lap_history_count;
 	uint8_t		in_pit;
 	uint8_t		pit_crossing_latched;
 	uint8_t		mandatory_pit_served;	/* count of 0x54
